// src/models/supplierBankDetailsDetail.model.ts
export interface SupplierBankDetailsDetailModel {
    bankName?: string;
    bankAccountCurrency?: string;
    effectiveDate?: string;
    bankAccountHolderName?: string;
    nameIsDifferentFromBankAccountName?: boolean;
    reasonName?: string;
    factoryCompany?: boolean;
    reasonFactory?: string;
    bankAccountNumber?: string;
    ibanNumber?: string;
    swiftCode?: string;
    sortCode?: string;
}